requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-bootstrap-form==3.4
idna==3.10
mysqlclient==2.2.7
pydantic==2.11.4
pydantic_core==2.33.2
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
stripe==12.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
